feat: solve day 14
this required using reddit for hints, as for some reason i couldn't get
past the brute force solution, despite how obvious the performant
solution is. probably partially due to my obsession to experiment with
parallellising work. after some time i had to concede there was no
solution down that road, as is so often the case.