feat(solutions): solve day 12
this solution reuses a Dijkstra's algorithm implementation from 2021,
which saved me a bit of work. i did need to rework it a bit, because
this year's problem required using both from and to coordinates for cost
calculation.

the second part is hugely wasteful, as the traversal cost calculations
are discarded after each starting point.

it should be possible to swap starting and ending points, and creating a
traversal cost map that could be reused. it sounds simple, but there are
some gotchas that i can already see are going to be tricky to tackle.

i might have a go at that later, just to see how much faster it would
be, and also just to see if i can make it work.