feat(solutions): solve day 13
very frustrating day today. started solving the problem in Go, but kept
failing with the actual data, despite the test data giving me the right
result.

eventually gave up on finding a parsing (or non-parsing) solution in Go
and pulled out NodeJS instead, as the data was JSON and the ruleset
fairly easily implemented in a dynamically typed language. i'm sure it
would be possible to replicate in Go, but nobody uses the reflections
package voluntarily, and i sort of feel like that would be the way to go.

despite not enjoying it, i still managed to get both stars after way too
many hours of head scratching and several failed attempts.