feat(solutions): (re)solve day 13
after having to resort to using NodeJS to crack the nut yesterday, i
browsed others' solutions and came across a really neat Go
[implementation by @dvicente@mas.to][1]. i more or less copied the parsing
code, which was the bit that got the better of me, so this find was very
edifying in showing me where my thinking in Go needs more work.

after that part was sorted, the rest was plain sailing, especially as i
had the right solutions for final data to lean on, too.

[1] https://mastodon.social/@dvicente@mas.to/109508026130023119